requirements.txt
jupyterlab
numpy
pandas
matplotlib
python-dotenv